+++ spec/fixtures/packages/package-that-throws-on-activate/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+  activate() { throw new Error('Top that'); },
+  deactivate() {},
+  serialize() {}
+};