+++ spec/fixtures/packages/package-with-provided-services/index.js
@@ -1,0 +1,26 @@
+/*
+ * decaffeinate suggestions:
+ * DS102: Remove unnecessary code created because of implicit returns
+ * Full docs: https://github.com/decaffeinate/decaffeinate/blob/main/docs/suggestions.md
+ */
+module.exports = {
+  activate() {},
+
+  deactivate() {},
+
+  provideFirstServiceV2() {
+    return 'first-service-v2';
+  },
+
+  provideFirstServiceV3() {
+    return 'first-service-v3';
+  },
+
+  provideFirstServiceV4() {
+    return 'first-service-v4';
+  },
+
+  provideSecondService() {
+    return 'second-service';
+  }
+};