+++ spec/fixtures/task-handler-with-deprecations.js
@@ -1,0 +1,4 @@
+const {deprecate} = require('grim');
+
+deprecate('Fake task deprecation');
+module.exports = function() {};